api/requirements.txt
azure-functions
pandas
pyarrow
adlfs
azure-storage-file-datalake
azure-identity
python-dotenv
